+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=C:\Desenvolvimento\framework\flutter"
-export "FLUTTER_APPLICATION_PATH=C:\Desenvolvimento\Source - Flutter\Aletheia - Project\source\core\aletheia_core_model"
+export "FLUTTER_APPLICATION_PATH=D:\Desenvolvimento\Source - Flutter\aletheia-core-model"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=0.0.1"
-export "FLUTTER_BUILD_NUMBER=0.0.1"
+export "FLUTTER_BUILD_NAME=0.0.3"
+export "FLUTTER_BUILD_NUMBER=0.0.3"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"